Use default AWS credentials provider
Previously if no access key was given in an AWS config I was using
InstanceProfileCredentialsProvider, which uses the instance metadata
(IMDS).  That doesn't work in Lambda, so I fell back to using
DefaultCredentialsProvider which checks about 5 things.  I'm kind of
curious which one Lambda is actually using, but it works.